Add files via upload
Welcome to my personal portfolio! This website showcases my professional journey, skills, and projects. Built entirely with HTML and CSS, it represents my ability to create clean, and visually appealing web pages. Below is a breakdown of the main sections of the portfolio:

Sections:
About Me:
-> An introduction to who I am and what I do.
-> Personal anecdotes or stories that highlight my passion for web development.

Skills:
-> A comprehensive list of my technical skills, including programming languages.

Projects:
-> A showcase of my key projects, each with:
-> A title and brief description.
-> Screenshots.
-> A list of technologies used.

Contact:
-> A contact form allowing visitors to send me messages directly from the website.
-> My email address and other contact information for easy reach.
-> Links to my social media profiles for further networking.

Features:
-> Clean and Modern Layout: The design emphasizes simplicity and readability, making use of modern web design principles.
-> Interactive Elements: Includes hover effects, smooth scrolling, and other interactive elements to enhance user engagement.
-> Accessibility: Ensures that the website is accessible to users with disabilities by adhering to web accessibility standards.

Technologies Used:
HTML5: Structured the content and provided semantic meaning to the web pages.
CSS3: Styled the website with a focus on layout, colors, and typography. Utilized Flexbox and Grid for responsive design.